move tests into commonTest, move HttpConnection into common code
